Moved themes to separate file
